+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - nois-drand: use unchecked `G2Pubkey` due to static pubkey
 - gateway: Reduce jobs per submission to 1 for verifying transactions and 10 for
   nonn-verifying transactions
+- nois-drand: Incentivized rounds are now those with jobs or `round % 25 == 0`.
 
 [#291]: https://github.com/noislabs/nois-contracts/pull/291